Implement user movement sharing
We considered 'user's ping is random'.
So we created a movement history queue, and we used to store the paths.

In server and client communications, movement has a delay of 0.1 seconds (0.1 is my setting value)
And the move history stores up to 10 location information. 
This can result in 0.1 to 2(0.1*10 * two shared clients) second movement delay depending on ping.

If the ping is too high and the Queue is full, the server and other clients will compress his(sender) movements.
So we can end up with a little bit of response time, and we can get a stable movement.

-> The packet is sent every 0.1 second, and between these 0.1 seconds we applied movement interpolation technology.