Implementing an Android Database for Post System
The client stores the received information  in a database.
Therefore, when user request post list, user can select only the data that you do not receive.

postal system (server - client)
When the new mail arrives, the server sends the zip code to the client.
The client sends the last postal code he received to the server.
The server can then only send mail that the client did not receive.